requirements.txt
nose
coverage
flake8-html
junit2html
antlr4-python3-runtime
